// packages/backend/src/types/schemas/prisma-generated/outputTypeSchemas/UserFindManyArgsSchema.ts
export const UserSelectSchema: z.ZodType<Prisma.UserSelect> = z.object({
  id: z.boolean().optional(),
  email: z.boolean().optional(),
  firebaseUid: z.boolean().optional(),
  name: z.boolean().optional(),
  role: z.boolean().optional(),
  planType: z.boolean().optional(),
  isAdmin: z.boolean().optional(),
  isVerified: z.boolean().optional(),
  verificationToken: z.boolean().optional(),
  resetPasswordToken: z.boolean().optional(),
  resetPasswordExpires: z.boolean().optional(),
  invitationExpires: z.boolean().optional(),
  invitedBy: z.boolean().optional(),
  createdAt: z.boolean().optional(),
  updatedAt: z.boolean().optional(),
  isFirstLogin: z.boolean().optional(),
  profile: z.union([z.boolean(),z.lazy(() => ProfileArgsSchema)]).optional(),
  projects: z.union([z.boolean(),z.lazy(() => ProjectFindManyArgsSchema)]).optional(),
  sentMessages: z.union([z.boolean(),z.lazy(() => MessageFindManyArgsSchema)]).optional(),
  receivedMessages: z.union([z.boolean(),z.lazy(() => MessageFindManyArgsSchema)]).optional(),
  bookmarks: z.union([z.boolean(),z.lazy(() => BookmarkFindManyArgsSchema)]).optional(),
  reports: z.union([z.boolean(),z.lazy(() => ReportFindManyArgsSchema)]).optional(),
  notifications: z.union([z.boolean(),z.lazy(() => NotificationFindManyArgsSchema)]).optional(),
  notificationArchives: z.union([z.boolean(),z.lazy(() => NotificationArchiveFindManyArgsSchema)]).optional(),
  notificationSetting: z.union([z.boolean(),z.lazy(() => NotificationSettingArgsSchema)]).optional(),
  subscription: z.union([z.boolean(),z.lazy(() => SubscriptionArgsSchema)]).optional(),
  pendingSubscription: z.union([z.boolean(),z.lazy(() => PendingSubscriptionArgsSchema)]).optional(),
  subscriptionHistory: z.union([z.boolean(),z.lazy(() => SubscriptionHistoryFindManyArgsSchema)]).optional(),
  activityLogs: z.union([z.boolean(),z.lazy(() => ActivityLogFindManyArgsSchema)]).optional(),
  payments: z.union([z.boolean(),z.lazy(() => PaymentFindManyArgsSchema)]).optional(),
  errorLogs: z.union([z.boolean(),z.lazy(() => ErrorLogFindManyArgsSchema)]).optional(),
  mentionedInMessages: z.union([z.boolean(),z.lazy(() => MessageFindManyArgsSchema)]).optional(),
  createdInvitations: z.union([z.boolean(),z.lazy(() => InvitationCodeFindManyArgsSchema)]).optional(),
  usedInvitations: z.union([z.boolean(),z.lazy(() => InvitationCodeFindManyArgsSchema)]).optional(),
  createdTemplates: z.union([z.boolean(),z.lazy(() => EmailTemplateFindManyArgsSchema)]).optional(),
  updatedTemplates: z.union([z.boolean(),z.lazy(() => EmailTemplateFindManyArgsSchema)]).optional(),
  sentEmails: z.union([z.boolean(),z.lazy(() => EmailLogFindManyArgsSchema)]).optional(),
  articles: z.union([z.boolean(),z.lazy(() => ArticleFindManyArgsSchema)]).optional(),
  _count: z.union([z.boolean(),z.lazy(() => UserCountOutputTypeArgsSchema)]).optional(),
}).strict()

export const UserFindManyArgsSchema: z.ZodType<Prisma.UserFindManyArgs> = z.object({
  select: UserSelectSchema.optional(),
  include: z.lazy(() => UserIncludeSchema).optional(),
  where: UserWhereInputSchema.optional(),
  orderBy: z.union([ UserOrderByWithRelationInputSchema.array(),UserOrderByWithRelationInputSchema ]).optional(),
  cursor: UserWhereUniqueInputSchema.optional(),
  take: z.number().optional(),
  skip: z.number().optional(),
  distinct: z.union([ UserScalarFieldEnumSchema,UserScalarFieldEnumSchema.array() ]).optional(),
}).strict() ;